+++ aggregate.rb
@@ -1,84 +1,105 @@
 require_relative 'config/initialize'
 require_relative 'db/connect'
-require_relative 'lib/denebola_logger'
 Dir['models/concern/*'].each {|f| require_relative f }
 Dir['models/*.rb'].each {|f| require_relative f }
 
-logger = DenebolaLogger.new(Settings.logger.path.aggregate)
+class Aggregator
+  include ArgumentUtil
 
-def extra_attribute(race, entry, horse)
-  entry_time = race.start_time
-  results_before = horse.results_before(entry_time)
+  def self.work!
+    logger = DenebolaLogger.new(Settings.logger.path.aggregate)
+    ArgumentUtil.logger = logger
 
-  blank = if results_before.second
-            (entry_time.to_date - results_before.second.race.start_time.to_date).to_i
-          else
-            0
-          end
+    check_operation(operation)
+    ApplicationRecord.operation = operation
 
-  sum_distance = results_before.map {|result| result.race.distance }.inject(:+)
-  average_distance = sum_distance / horse.entry_times(entry_time).to_f
-  distance_diff = (race.distance - average_distance).abs / average_distance
+    aggregator = new(logger)
 
-  {
-    blank: blank,
-    distance_diff: distance_diff,
-    entry_times: horse.entry_times(entry_time),
-    horse_average_prize_money: horse.average_prize_money(entry_time),
-    jockey_average_prize_money: entry.jockey.average_prize_money(entry_time),
-    jockey_win_rate: entry.jockey.win_rate(entry_time),
-    jockey_win_rate_last_four_races: entry.jockey.win_rate_last_four_races(entry_time),
-    last_race_order: horse.last_race_order(entry_time),
-    month: race.month,
-    rate_within_third: horse.rate_within_third(entry_time),
-    second_last_race_order: horse.second_last_race_order(entry_time),
-    weight_per: entry.weight_per,
-    win_times: horse.win_times(entry_time),
-    won: entry.won,
-  }
-end
+    logger.info('Start Aggregation')
 
-logger.info('Start Aggregation')
+    entries = Entry.joins(:race).joins(:horse)
+                   .where(order: (1..18).to_a.map(&:to_s))
+                   .where.not(weight: nil)
+                   .where('DATE(entries.updated_at) >= ?', from.strftime('%F'))
+                   .where('DATE(entries.updated_at) <= ?', to.strftime('%F'))
+                   .uniq
+    logger.info("# of Target Features = #{entries.size}")
 
-entries = Entry.joins(:race).joins(:horse)
-               .where(order: (1..18).to_a.map(&:to_s))
-               .where.not(weight: nil)
-               .pluck('races.race_id', 'horses.horse_id').uniq
-entries.map! {|race_id, horse_id| [race_id.to_s, horse_id.to_s] }
-features = Feature.pluck(:race_id, :horse_id).uniq
-new_features = entries - features
+    entries.each {|entry| aggregator.create_feature(entry) }
 
-logger.info("# of Updated Features = #{new_features.size}")
+    logger.info('Finish Aggregation')
+  end
 
-new_features.each do |race_id, horse_id|
-  attribute = {race_id: race_id, horse_id: horse_id}
-  feature_attributes = Feature.attribute_names - %w[horse_id race_id]
+  def initialize(logger)
+    @logger = logger
+  end
 
-  race = Race.find_by(race_id: race_id)
-  next unless race
+  def create_feature(entry)
+    attribute = create_feature_attribute(entry)
+    return if attribute.nil?
 
-  attribute.merge!(race.attributes.slice(*feature_attributes)).symbolize_keys!
+    begin
+      feature = Feature.create_or_update!(attribute)
+      @logger.info(base_log_attribute.merge(feature_id: feature.id))
+    rescue ActiveRecord::RecordInvalid => e
+      @logger.error(base_log_attribute.merge(errors: e.record.errors))
+      raise
+    end
+  end
 
-  horse = Horse.find_by(horse_id: horse_id)
-  next unless horse
+  private
 
-  attribute.merge!(horse.attributes.slice(*feature_attributes)).symbolize_keys!
+  def base_log_attribute
+    @base_log_attribute ||= {action: Feature.operation, resource: 'feature'}
+  end
 
-  entry = Entry.find_by(race_id: race.id, horse_id: horse.id)
-  next unless entry&.jockey
+  def create_feature_attribute(entry)
+    return if entry.race.nil? or entry.horse.nil? or entry.jockey.nil?
 
-  attribute.merge!(entry.attributes.slice(*feature_attributes)).symbolize_keys!
+    attribute = {race_id: entry.race.race_id, horse_id: entry.horse.horse_id}
+    feature_attribute_names = Feature.attribute_names - %w[horse_id race_id]
 
-  attribute.merge!(extra_attribute(race, entry, horse))
+    attribute.merge!(entry.race.attributes.slice(*feature_attribute_names))
+    attribute.merge!(entry.horse.attributes.slice(*feature_attribute_names))
+    attribute.merge!(entry.attributes.slice(*feature_attribute_names))
+    attribute.merge(extra_attribute(entry)).symbolize_keys
+  end
 
-  base_log_attribute = {action: 'create', resource: 'feature'}
-  begin
-    feature = Feature.create!(attribute.except(:id, :order))
-    logger.info(base_log_attribute.merge(feature_id: feature.id))
-  rescue ActiveRecord::RecordInvalid => e
-    logger.error(base_log_attribute.merge(errors: e.record.errors))
-    raise
+  def extra_attribute(entry)
+    race = entry.race
+    horse = entry.horse
+    jockey = entry.jockey
+
+    entry_time = race.start_time
+    results_before = horse.results_before(entry_time)
+
+    blank = if results_before.second
+              (entry_time.to_date - results_before.second.race.start_time.to_date).to_i
+            else
+              0
+            end
+
+    sum_distance = results_before.map {|result| result.race.distance }.inject(:+)
+    average_distance = sum_distance / horse.entry_times(entry_time).to_f
+    distance_diff = (race.distance - average_distance).abs / average_distance
+
+    {
+      blank: blank,
+      distance_diff: distance_diff,
+      entry_times: horse.entry_times(entry_time),
+      horse_average_prize_money: horse.average_prize_money(entry_time),
+      jockey_average_prize_money: jockey.average_prize_money(entry_time),
+      jockey_win_rate: jockey.win_rate(entry_time),
+      jockey_win_rate_last_four_races: jockey.win_rate_last_four_races(entry_time),
+      last_race_order: horse.last_race_order(entry_time),
+      month: race.month,
+      rate_within_third: horse.rate_within_third(entry_time),
+      second_last_race_order: horse.second_last_race_order(entry_time),
+      weight_per: entry.weight_per,
+      win_times: horse.win_times(entry_time),
+      won: entry.won,
+    }
   end
 end
 
-logger.info('Finish Aggregation')
+Aggregator.work!