+++ models/race.rb
@@ -1,6 +1,8 @@
 # coding: utf-8
 
-class Race < ActiveRecord::Base
+require_relative 'application_record'
+
+class Race < ApplicationRecord
   DIRECTION_LIST = %w[左 右 直 障].freeze
   GRADE_LIST = %w[G1 G2 G3 G J.G1 J.G2 J.G3 L OP].freeze
   PLACE_LIST = %w[中京 中山 京都 函館 小倉 新潟 札幌 東京 福島 阪神].freeze
@@ -19,7 +21,8 @@ class Race < ActiveRecord::Base
   has_one :trio
   has_one :trifecta
 
-  validates :distance, :direction, :place, :race_id, :round, :track, :weather,
+  validates :distance, :direction, :place, :race_id, :race_name, :round, :track,
+            :weather,
             presence: {message: 'absent'}
   validates :direction,
             inclusion: {in: DIRECTION_LIST, message: 'invalid'},
@@ -46,27 +49,30 @@ class Race < ActiveRecord::Base
             inclusion: {in: WEATHER_LIST, message: 'invalid'},
             allow_nil: true
 
+  def self.create_or_update!(attribute)
+    race = find_by(attribute.slice(:race_id))
+    super(race, attribute)
+  end
+
+  def self.log_attribute
+    super.merge(resource: 'race')
+  end
+
   def month
     start_time.month
   end
 
-  def create_payoff(attribute)
-    create_one_pattern_ticket(attribute.slice(*ONE_PATTERN_TICKET_LIST))
-    create_multi_patterns_ticket(attribute.slice(*MULTI_PATTERNS_TICKET_LIST))
-  end
-
-  private
-
-  def create_one_pattern_ticket(attribute)
-    attribute.each do |betting_ticket, attr|
-      send("create_#{betting_ticket}", attr) unless send(betting_ticket)
+  def create_or_update_payoff(attribute)
+    attribute.slice(*ONE_PATTERN_TICKET_LIST).each do |betting_ticket, attr|
+      ticket = send(betting_ticket)
+      ticket.update!(attr) if ticket.present? and self.class.updatable?
+      send("create_#{betting_ticket}", attr) if ticket.nil? and self.class.creatable?
     end
-  end
 
-  def create_multi_patterns_ticket(attribute)
-    attribute.each do |betting_ticket, attrs|
+    attribute.slice(*MULTI_PATTERNS_TICKET_LIST).each do |betting_ticket, attrs|
       attrs.each do |attr|
-        send(betting_ticket).find_or_create_by!(attr)
+        tickets = send(betting_ticket)
+        tickets.create!(attr) unless tickets.exists?(attr.except(:odds))
       end
     end
   end