+++ spec/models/race_spec.rb
@@ -11,10 +11,12 @@ describe Race, type: :model do
         grade: ['G1', 'G2', 'G3', 'G', 'J.G1', 'J.G2', 'J.G3', 'L', 'OP', nil],
         place: %w[中京 中山 京都 函館 小倉 新潟 札幌 東京 福島 阪神],
         race_id: %w[0],
+        race_name: %w[レース名],
         round: [1],
         track: %w[芝 ダート 障],
         weather: %w[晴 曇 小雨 雨 小雪 雪],
       }
+      include_context 'トランザクション作成'
       it_behaves_like '正常な値を指定した場合のテスト', valid_attribute
     end
 
@@ -29,7 +31,8 @@ describe Race, type: :model do
         track: ['invalid'],
         weather: ['invalid'],
       }
-      absent_keys = invalid_attribute.keys - %i[grade]
+      absent_keys = invalid_attribute.keys - %i[grade] + %i[race_name]
+      include_context 'トランザクション作成'
       it_behaves_like '必須パラメーターがない場合のテスト', absent_keys
       it_behaves_like '不正な値を指定した場合のテスト', invalid_attribute
     end
@@ -39,8 +42,10 @@ describe Race, type: :model do
     describe '正常系' do
       (1..12).each do |i|
         context "#{i}月のレースの場合" do
+          include_context 'トランザクション作成'
           before(:all) do
-            start_time = Time.parse("2000/#{format('%02d', i)}/01 00:00:00")
+            time_string = "2000/#{format('%<month>02d', month: i)}/01 00:00:00"
+            start_time = Time.parse(time_string)
             @race = build(:race, start_time: start_time)
           end
 
@@ -49,6 +54,21 @@ describe Race, type: :model do
           end
         end
       end
+    end
+  end
+
+  describe '.create_or_update!' do
+    describe '正常系' do
+      it_behaves_like '.create_or_update!: データが既に存在する場合のテスト',
+                      {'round' => 2},
+                      %i[race_id]
+      it_behaves_like '.create_or_update!: データが存在しない場合のテスト'
+    end
+  end
+
+  describe '.log_attribute' do
+    describe '正常系' do
+      it_behaves_like '.log_attribute: 返り値が正しいこと'
     end
   end
 end