+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 2.8.2 (2019/12/08)
+- [REFACTOR] collector, extractor and aggregator
+
+# 2.8.1 (2019/12/02)
+- [UPDATE] ruby version to 2.6.3
+
+# 2.8.0 (2019/11/29)
+- [UPDATE] script to add option for operation
+- [UPDATE] gems
+
+# 2.7.0 (2019/11/19)
+- [ADD] new race attribute
+- [UPDATE] gems
+
 # 2.6.2 (2019/09/23)
 - [FIX] extraction script
 - [UPDATE] index on entries table